+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Fixed
 - `getSearchContent()` was incorrectly spelled `gtSearchContent()` in the `CanBeSearched` trait.
+- If the config does not contain a 'models' key, we will make sure we default to an array in the `IndexCommand`.
 
 ## [0.1.1] - 2017-06-04