+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [UNRELEASED]
 
+## [1.0.0] 2017-10-30
+
 ### Added
 - Queue tubes are now customizable.
 - Support for Laravel auto-discovery.
@@ -30,5 +32,6 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Added
 - New package :)
 
-[unreleased]: https://github.com/browner12/larasearch/compare/v0.1.1...HEAD
+[unreleased]: https://github.com/browner12/larasearch/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/browner12/larasearch/compare/v0.1.1...v1.0.0
 [0.1.1]: https://github.com/browner12/larasearch/compare/v0.1.0...v0.1.1